+++ front/src/components/organisms/index.ts
@@ -17,6 +17,7 @@ import type { TResourceForm } from './ResourceForm'
 import { UserAccessHome } from './UserAccessHome'
 import { UserProfileResourcesWidget } from './UserProfileResourcesWidget'
 import { VotesDateController } from './VotesDateController'
+import { Search } from './Search'
 
 export {
   AccountAdmin,
@@ -37,6 +38,7 @@ export {
   UserAccessHome,
   UserProfileResourcesWidget,
   VotesDateController,
+  Search,
 }
 
 export type { TResourceForm }