+++ front/src/types/types.tsx
@@ -97,6 +97,7 @@ export type TFilters = {
   resourceTypes?: string[]
   status?: string[]
   topic?: string
+  search?: string
 }
 export type TEditResourceProps = {
   description: string